requirements.txt
openai-whisper
google-generativeai
moviepy
mediapipe
streamlit
Pillow
numpy
